feat: add mongoDB id validation

+++ src/app/Collection/collectionController.js
@@ -2,6 +2,7 @@ const collectionProvider = require('./collectionProvider')
 const collectionService = require('./collectionService')
 const { response, errResponse } = require("../../../config/response");
 const baseResponseStatus = require('../../../config/baseResponseStatus');
+const mongoose = require("mongoose");
 
 exports.postCollection = async(req, res) => {
   const { name } = req.body
@@ -16,8 +17,11 @@ exports.postCollection = async(req, res) => {
 exports.postPlace = async (req, res) =>{
   const collectionId = req.params.collectionId;
   const { placeId } = req.body
+  if(!placeId) return res.send(errResponse(baseResponseStatus.PLACE_ID_EMPTY))
+  if(!mongoose.isValidObjectId(placeId)) return res.send(errResponse(baseResponseStatus.PLACE_ID_WRONG))
   const collection = collectionId.trim()
   if(!collection || collection === ':collectionId') return res.send(errResponse(baseResponseStatus.COLLECTION_ID_EMPTY))
+  if(!mongoose.isValidObjectId(collection)) return res.send(errResponse(baseResponseStatus.COLLECTION_ID_WRONG))
   if(!placeId || placeId.length < 1) return res.send(errResponse(baseResponseStatus.PLACE_ID_EMPTY))
   const pushPlaceRes = await collectionService.pushPlace(collection, placeId.trim())
   return res.send(pushPlaceRes)
@@ -33,6 +37,7 @@ exports.getCollection = async (req, res) => {
   const collectionId = req.params.collectionId;
   const collection = collectionId.trim()
   if(!collection || collection === ':collectionId') return res.send(errResponse(baseResponseStatus.COLLECTION_ID_EMPTY))
+  if(!mongoose.isValidObjectId(collection)) return res.send(errResponse(baseResponseStatus.COLLECTION_ID_WRONG))
   const getCollectionRes = await collectionProvider.getCollection(collection)
   return res.send(getCollectionRes)
 }
@@ -40,9 +45,11 @@ exports.getCollection = async (req, res) => {
 exports.updateCollectionName = async (req, res) => {
   const collectionId = req.params.collectionId;
   const { collectionName } = req.body
+  if (!collectionName) return res.send(errResponse(baseResponseStatus.COLLECTION_UPDATE_NAME_EMPTY))
   const collection = collectionId.trim()
   const name = collectionName.trim()
   if(!collection || collection === ':collectionId') return res.send(errResponse(baseResponseStatus.COLLECTION_ID_EMPTY))
+  if(!mongoose.isValidObjectId(collection)) return res.send(errResponse(baseResponseStatus.COLLECTION_ID_WRONG))
   if(!name || name.length < 1) return res.send(errResponse(baseResponseStatus.COLLECTION_UPDATE_NAME_EMPTY))
   const updateCollectionRes = await collectionService.patchCollectionName(collection, name)
   return res.send(updateCollectionRes)
@@ -53,7 +60,9 @@ exports.deletePlaceInCollection = async (req, res) => {
   const collection = collectionId.trim()
   const place = placeId.trim()
   if(!collection || collection === ':collectionId') return res.send(errResponse(baseResponseStatus.COLLECTION_ID_EMPTY))
+  if(!mongoose.isValidObjectId(collection)) return res.send(errResponse(baseResponseStatus.COLLECTION_ID_WRONG))
   if(!place || place === ':placeId') return res.send(errResponse(baseResponseStatus.PLACE_ID_EMPTY))
+  if(!mongoose.isValidObjectId(place)) return res.send(errResponse(baseResponseStatus.PLACE_ID_WRONG))
   const deletePlaceRes = await collectionService.deletePlaceInCollection(collection, place)
   return res.send(deletePlaceRes)
 }
@@ -62,6 +71,7 @@ exports.deleteCollection = async (req, res) => {
   const collectionId = req.params.collectionId;
   const collection = collectionId.trim()
   if(!collection || collection === ':collectionId') return res.send(errResponse(baseResponseStatus.COLLECTION_ID_EMPTY))
+  if(!mongoose.isValidObjectId(collection)) return res.send(errResponse(baseResponseStatus.COLLECTION_ID_WRONG))
   const deletedRes = await collectionService.deleteCollection(collection)
   return res.send(deletedRes)
 }
